requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.2.0
paho-mqtt==2.1.0
propcache==0.3.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
telegram==0.0.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3